integrated ML Rule to be added to consumer stream

+++ src/main/java/com/example/fraud/event/FraudResultEvent.java
@@ -11,6 +11,6 @@ public class FraudResultEvent {
     private String transactionId;
     private String decision;      // "FRAUD" or "APPROVED"
     private String reason;        // e.g. "AMOUNT_EXCEEDS_LIMIT"
-    private Double score;         // optional for later ML
+    private Float score;         // optional for later ML
     private Instant evaluatedAt;
 }
